third_party/WebKit/LayoutTests/media/track/track-cue-rendering-wider-than-controls-expected.txt

Test that the cue root is not constrained by the controls/overlay.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (parseFloat(getComputedStyle(controlsContainer).width) <= '800') OK
EXPECTED (parseFloat(getComputedStyle(cueRoot).width) > '800') OK
END OF TEST
